Extract import/export to class.
